+++ src/mock/services/CheerioServiceMock.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { Page } from '../domains/Page';
+import { Page } from '../../domains/Page';
 
 export class CheerioServiceMock {
   async retrieveDataByUrl(url: string): Promise<Partial<Page>> {
